DefID type is set to u16

+++ compiler-rt/lib/dfisan/dfisan_rtl.cpp
@@ -20,12 +20,12 @@
 using namespace __sanitizer;
 
 // ------------- Runtime Definition Table Management -----------
-static inline void setRDT(uptr Addr, u32 ID) {
+static inline void setRDT(uptr Addr, u16 ID) {
   u16 *shadow_memory = (u16 *)__dfisan::MemToShadow(Addr);
   *shadow_memory = ID;
   Report("SET: %d at %p\n", ID, (void *)shadow_memory);
 }
-static inline bool checkRDT(uptr Addr, u32 ID) {
+static inline bool checkRDT(uptr Addr, u16 ID) {
   u16 *shadow_memory = (u16 *)__dfisan::MemToShadow(Addr);
   Report("CHECK: %d at %p\n", ID, (void *)shadow_memory);
   return *shadow_memory == ID;
@@ -38,19 +38,20 @@ bool dfisan_inited = false;
 bool dfisan_init_is_running = false;
 
 extern "C" SANITIZER_INTERFACE_ATTRIBUTE
-void __dfisan_store_id(uptr StoreAddr, u32 DefID) {
+void __dfisan_store_id(uptr StoreAddr, u16 DefID) {
   Report("INFO: Set DefID(%d) at %p\n", DefID, (void *)StoreAddr);
   setRDT(StoreAddr, DefID);
 }
 
+// TODO: va_arg cannot use `u16` (these values are converted to int)
 extern "C" SANITIZER_INTERFACE_ATTRIBUTE
-void __dfisan_check_ids(uptr LoadAddr, u32 Argc, ...) {
+void __dfisan_check_ids(uptr LoadAddr, u16 Argc, ...) {
   Report("INFO: Check SetID at %p\n", (void *)LoadAddr);
   va_list Args;
   va_start(Args, Argc);
   bool NoErr = false;
-  for (u32 i = 0; i < Argc; i++) {
-    u32 ID = va_arg(Args, u32);
+  for (u16 i = 0; i < Argc; i++) {
+    u16 ID = va_arg(Args, u16);
     Report("Checking SetID(%d)...\n", ID);
     NoErr |= checkRDT(LoadAddr, ID);
   }
